rename RoutineSequence
