+++ src/main/java/com/sacristan/api/global/entities/assignments/routineSegment/RoutineSegmentRepository.java
@@ -1,4 +1,4 @@
-package com.sacristan.api.global.entities.assignments.routineSequence;
+package com.sacristan.api.global.entities.assignments.routineSegment;
 
 import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoutineSequenceRepository extends JpaRepository<RoutineSequence, Long> {
+public interface RoutineSegmentRepository extends JpaRepository<RoutineSegment, Long> {
 
     @Modifying
     @Transactional
